Act010_Main.java Act010_Main_Presenter.java ticket_creation_dialog.xml     - Adição de serviço de serial e dialog de criação de ticket. @#$ act010_main_content.xml     - Remoção de botão de serviço mockado. @#$ Sql_Act010_001.java     - Adaptação de query para listar ticket type com os forms @#$ Sql_Act010_002.java     -  query para listar ticket type @#$ TicketCreationParams.kt     - Ajuste de tipo. @#$ WSTicketCreation.kt     - Serviço de criação de tickets. @#$

+++ app/src/main/java/com/namoadigital/prj001/model/TicketCreationParams.kt
@@ -8,7 +8,7 @@ class TicketCreationParams(
     @SerializedName("site_code") var siteCode: Int,
     @SerializedName("operation_code") var operationCode: Long,
     @SerializedName("product_code") var productCode: Long,
-    @SerializedName("serial_code") var serialCode: Int,
+    @SerializedName("serial_code") var serialCode: Long,
     @SerializedName("comments") var comments: String
 ) {
